made some cleanup changes to my seeding files so I'm making a commit now in case I broke something, so that I can compare with the prior commit
